requirements.txt
flask
requests
pandas
flask-login
python-dotenv
